+++ common/src/main/java/com/jaquadro/minecraft/storagedrawers/block/framed/BlockFramedCompDrawers.java
@@ -6,12 +6,14 @@ import com.jaquadro.minecraft.storagedrawers.api.framing.IFramedBlock;
 import com.jaquadro.minecraft.storagedrawers.api.framing.IFramedBlockEntity;
 import com.jaquadro.minecraft.storagedrawers.block.BlockCompDrawers;
 import com.jaquadro.minecraft.storagedrawers.block.tile.*;
+import com.jaquadro.minecraft.storagedrawers.block.tile.tiledata.MaterialData;
 import com.jaquadro.minecraft.storagedrawers.components.item.FrameData;
 import com.jaquadro.minecraft.storagedrawers.core.ModDataComponents;
 import com.jaquadro.minecraft.storagedrawers.util.WorldUtils;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.LevelReader;
 import net.minecraft.world.level.block.state.BlockState;
@@ -74,5 +76,20 @@ public class BlockFramedCompDrawers extends BlockCompDrawers implements IFramedB
     @Override
     public boolean supportsFrameMaterial (FrameMaterial material) {
         return true;
+    }
+
+    protected float getShadeBrightness(BlockState state, BlockGetter level, BlockPos pos) {
+        if (isHalfDepth())
+            return 1f;
+
+        BlockEntityDrawersComp tile = WorldUtils.getBlockEntity(level, pos, BlockEntityDrawersComp.class);
+        if (tile == null)
+            return 1f;
+
+        MaterialData data = tile.material();
+        if (data == null)
+            return 1f;
+
+        return data.allMatOpaque() ? 0.8f : 1f;
     }
 }